requirements.txt
appdirs==1.4.4
apyori==1.1.2
asgiref==3.8.1
bcrypt==3.1.7
blinker==1.8.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cryptography>=41.0.5,<43
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
Django==4.2.19
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Fabric3==1.14.post1
filelock==3.15.4
flask==3.0.3
Flask-Cors==4.0.1
flask-sqlalchemy==3.1.1
fonttools==4.54.1
greenlet==3.0.3
gunicorn==23.0.0
importlib-metadata==8.0.0
importlib-resources==6.4.5
invoke==2.2.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mlxtend==0.23.1
mysqlclient==2.2.4
numpy==1.24.4
packaging==24.1
pandas==2.0.3
paramiko==2.12.0
pillow==10.4.0
platformdirs==4.2.2
pycodestyle==2.12.0
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.3.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.14.0
SQLAlchemy==2.0.31
sqlparse==0.5.3
threadpoolctl==3.5.0
typing-extensions==4.12.2
tzdata==2024.2
virtualenv==20.26.3
werkzeug==3.0.3
whitenoise==6.7.0
wrapt==1.16.0
zipp==3.19.2
